MOve profile guard to shared module
