+++ client/app/src/app/app-routing.module.ts
@@ -3,8 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import {
   RedirectToDefaultRouteComponent
 } from './shared/_components/redirect-to-default-route/redirect-to-default-route.component';
-import { CreateProfileGuardService } from './user-profile/create-profile-guard.service';
-
+import { CreateProfileGuardService } from './shared/_guards/create-profile-guard.service';
 import { LoginGuard } from './shared/_guards/login-guard.service';
 
 const routes: Routes = [
