+++ client/app/src/app/case/case-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule, Injectable } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { CreateProfileGuardService } from '../user-profile/create-profile-guard.service';
+import { CreateProfileGuardService } from '../shared/_guards/create-profile-guard.service';
 import { LoginGuard } from '../shared/_guards/login-guard.service';
 import { _TRANSLATE } from '../shared/translation-marker';
 import { CasesComponent } from './components/cases/cases.component';
