+++ client/app/src/app/class/class-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import {CreateProfileGuardService} from "../user-profile/create-profile-guard.service";
+import {CreateProfileGuardService} from "../shared/_guards/create-profile-guard.service";
 import {LoginGuard} from "../shared/_guards/login-guard.service";
 import {DashboardComponent} from "./dashboard/dashboard.component";
 import {ClassFormsPlayerComponent} from "./class-forms-player/class-forms-player.component";
