+++ client/app/src/app/core/export-data/export-data-routing.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { ExportDataComponent } from './export-data/export-data.component';
 import { LoginGuard } from '../../shared/_guards/login-guard.service';
-import { CreateProfileGuardService } from '../../user-profile/create-profile-guard.service';
+import { CreateProfileGuardService } from '../../shared/_guards/create-profile-guard.service';
 const routes = [{
   path: 'export-data',
   component: ExportDataComponent,
