+++ client/app/src/app/core/sync-records/sync-records-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { CreateProfileGuardService } from '../../user-profile/create-profile-guard.service';
+import { CreateProfileGuardService } from '../../shared/_guards/create-profile-guard.service';
 
 import { LoginGuard } from '../../shared/_guards/login-guard.service';
 import { SyncRecordsComponent } from './sync-records/sync-records.component';
